Replace unsigned long to uint32_t in mt19937ar

+++ source/mt19937ar/mt19937ar.h
@@ -43,17 +43,19 @@
    email: m-mat @ math.sci.hiroshima-u.ac.jp (remove space)
 */
 
+#include <stdint.h>
+
 /* initializes mt[N] with a seed */
-void init_genrand(unsigned long s);
+void init_genrand(uint32_t s);
 
 /* initialize by an array with array-length */
 /* init_key is the array for initializing keys */
 /* key_length is its length */
 /* slight change for C++, 2004/2/26 */
-void init_by_array(unsigned long init_key[], int key_length);
+void init_by_array(uint32_t init_key[], int key_length);
 
 /* generates a random number on [0,0xffffffff]-interval */
-unsigned long genrand_int32(void);
+uint32_t genrand_int32(void);
 
 /* generates a random number on [0,0x7fffffff]-interval */
 long genrand_int31(void);
